refactor: Abstract class를 만들고 이를 상속받아 test 코드 중복 삭제
