+++ src/test/java/domain/blackjack/TestPlayerCardDrawStrategy.java
@@ -1,11 +1,6 @@
 package domain.blackjack;
 
-import domain.blackjack.Gamer;
-import domain.card.Card;
-import domain.card.CardDrawStrategy;
-import java.util.List;
-
-public class TestPlayerCardDrawStrategy implements CardDrawStrategy {
+public class TestPlayerCardDrawStrategy extends TestAbstractFirstCardDrawStrategy {
     private final Gamer player;
 
     public TestPlayerCardDrawStrategy(Gamer player) {
@@ -13,18 +8,7 @@ public class TestPlayerCardDrawStrategy implements CardDrawStrategy {
     }
 
     @Override
-    public final Card nextCard(List<Card> cards) {
-        if (canDraw()) {
-            return cardSelectStrategy(cards);
-        }
-        throw new IllegalStateException("카드를 더이상 뽑을 수 없습니다.");
-    }
-
-    private boolean canDraw() {
+    boolean canDraw() {
         return !player.getSummationCardPoint().isDeadPoint();
-    }
-
-    private Card cardSelectStrategy(List<Card> cards) {
-        return cards.get(0);
     }
 }
